test(http): Move event trigger test from `ApplicationTest` to `ApplicationCoreTest` for better organization.
